# apps/api/openapi.gen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anarlog API",
    "description": "Anarlog cloud services and opt-in hosted meeting access",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/nango/webhook": {
      "post": {
        "tags": [
          "nango"
        ],
        "operationId": "nango_webhook",
        "responses": {
          "200": {
            "description": "Webhook processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Invalid signature"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "WebhookResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token"
      },
      "cloud_api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "anl_...",
        "description": "Anarlog cloud API key"
      }
    }
  },
...
}